remove deupication of skip_unsupported_pc_sampling_soc
